refactor(date-field): organize component architecture
