+++ packages/admin-ui/src/date-field/date-field.style.ts
@@ -2,9 +2,20 @@ import { style } from '@vtex/admin-ui-core'
 
 export const dateField = style({
   display: 'flex',
+  justifyContent: 'space-between',
+  alignContent: 'center',
+  padding: '$s',
+  height: 52,
+  width: 288,
+  border: '$neutral',
+  borderRadius: 4,
 })
 
-export const dateFieldSegment = style({
+export const segmentList = style({
+  display: 'flex',
+})
+
+export const segment = style({
   cursor: 'pointer',
   text: '$body',
   bg: '$action.neutral.tertiary',
@@ -22,4 +33,9 @@ export const dateFieldSegment = style({
     color: '$action.main.primary',
     outline: 'none',
   },
+})
+
+export const label = style({
+  text: '$detail',
+  color: '$secondary',
 })
