+++ packages/admin-ui/src/date-field/index.ts
@@ -1,3 +1,2 @@
 export * from './date-field'
 export * from './date-field-state'
-export * from './date-field-segment'
