+++ packages/admin-ui/src/date-field/segment-list.ts
@@ -1,10 +1,11 @@
 import { Composite } from 'reakit'
 import { createComponent, useElement } from '@vtex/admin-ui-react'
+
 import * as style from './date-field.style'
 
-export const DateField = createComponent<typeof Composite>((props) =>
+export const SegmentList = createComponent<typeof Composite>((props) =>
   useElement(Composite, {
-    baseStyle: style.dateField,
+    baseStyle: style.segmentList,
     ...props,
   })
 )
