+++ packages/admin-ui/src/date-picker/date-picker.tsx
@@ -1,46 +1,44 @@
 import React from 'react'
 import { IconCalendar } from '@vtex/phosphor-icons'
 import { ariaAttr } from '@vtex/admin-ui-util'
-import { Flex } from '../components/Flex'
-import { Label } from '../components/Label'
 
 import { DateField } from '../date-field'
-import { DatePickerSegment } from './date-picker-segment'
 import { Picker, PickerDisclosure, PickerPopover } from '../picker'
 import { Calendar } from '../calendar'
 import type { DatePickerStateReturn } from './date-picker-state'
 import * as style from './date-picker.style'
 
 export function DatePicker(props: DatePickerProps) {
-  const { state, label } = props
+  const {
+    state: {
+      validationState,
+      isRequired,
+      pickerState,
+      dateFieldState,
+      calendarState,
+    },
+    label,
+  } = props
 
   return (
     <>
       <Picker
-        csx={style.datePicker}
-        aria-invalid={ariaAttr(state.validationState === 'invalid')}
-        aria-required={ariaAttr(state.isRequired)}
-        state={state.pickerState}
+        aria-invalid={ariaAttr(validationState === 'invalid')}
+        aria-required={ariaAttr(isRequired)}
+        state={pickerState}
       >
-        <Flex align="center" justify="space-between">
-          <Flex direction="column">
-            <Label csx={style.datePickerLabel}>{label}</Label>
-            <DateField state={state.dateFieldState}>
-              {state.dateFieldState.segments.map((segment, i) => (
-                <DatePickerSegment key={i} segment={segment} state={state} />
-              ))}
-            </DateField>
-          </Flex>
-          <PickerDisclosure
-            state={state.pickerState}
-            csx={style.datePickerDisclosure}
-          >
-            <IconCalendar />
-          </PickerDisclosure>
-        </Flex>
+        <DateField
+          label={label}
+          state={dateFieldState}
+          disclosure={
+            <PickerDisclosure state={pickerState} csx={style.disclosure}>
+              <IconCalendar />
+            </PickerDisclosure>
+          }
+        />
       </Picker>
-      <PickerPopover state={state.pickerState}>
-        <Calendar state={state.calendarState} />
+      <PickerPopover state={pickerState}>
+        <Calendar state={calendarState} />
       </PickerPopover>
     </>
   )
